requirements.txt
absl-py==0.12.0
astor==0.8.1
cached-property==1.5.2
click==7.1.2
cycler==0.10.0
dataclasses==0.8
fastapi==0.63.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.36.1
h11==0.12.0
h5py==3.1.0
importlib-metadata==3.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.5.1.48
opt-einsum==3.3.0
Pillow==8.2.0
protobuf==3.15.7
pydantic==1.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-multipart==0.0.5
six==1.15.0
starlette==0.13.6
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
typing-extensions==3.7.4.3
uvicorn==0.13.4
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
